requirements-dev.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tweepy==4.12.1
urllib3==1.26.12
